Revisi version Hero
